xsync: secure against unsafe /proc/$SLAVEPID/* access
Only the ssh connection setting up the idler connects with
the slave hostname. Termination of the $SLAVEPID process
at the slave site results in immediate termination of
sshd. So as long as all other connections, which refer to
/proc/$SLAVEPID strictly inherit the master connection,
we should be safe against bad /proc/$SLAVEPID/* access.

Unfortunately ssh does not support a strict inhertance
(of master connection) mode. But if it does inherit,
it does not bother resolving the hostname provided.

By exploiting the above described behavior, the scheme
is to set up the "master" connection with a valid hostname,
and all child ssh connections use "IDLER.IS.DEAD" as
the hostname. This way if the $SLAVEPID is dead, child ssh
gets forced to resolved "IDLER.IS.DEAD" and fail.

Signed-off-by: Anand Avati <avati@redhat.com>